backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pydantic>=2.10.0
pymupdf>=1.24.0
jsonschema>=4.23.0
pytesseract>=0.3.13
Pillow>=10.4.0
python-multipart>=0.0.9